refactor(ai): extract model + copilot DTOs into vailabel-models/vailabel-copilot
Phase 1 step 8d (final extraction). Moves the pure, self-contained AI request
DTOs out of domain/ai/model.rs:
- vailabel-models/contracts: ModelImport/Install/Component, GitHubReleaseLookup,
  ModelActivation, RuntimeInstall payloads.
- vailabel-copilot/contracts: CopilotTurn/Action/Test payloads + CopilotTestResult.
ai/model.rs re-exports them and keeps the inference/plugin-coupled types
(PipelineRunPayload<-PromptInput, InferencePoint/Draft) and the Rust-built
CopilotLlmConfig in the binary. Completes Phase 1 module-crate extraction.
