+++ apps/studio/src-tauri/Cargo.toml
@@ -15,6 +15,8 @@ members = [
     "crates/dataset",
     "crates/analysis",
     "crates/video",
+    "crates/models",
+    "crates/copilot",
     "crates/annotation",
     "crates/search",
     "crates/training",
@@ -39,6 +41,8 @@ vailabel-project = { path = "crates/project" }
 vailabel-dataset = { path = "crates/dataset" }
 vailabel-analysis = { path = "crates/analysis" }
 vailabel-video = { path = "crates/video" }
+vailabel-models = { path = "crates/models" }
+vailabel-copilot = { path = "crates/copilot" }
 vailabel-plugin = { path = "crates/plugin" }
 vailabel-runtime = { path = "crates/runtime" }
 vailabel-annotation = { path = "crates/annotation" }
@@ -83,6 +87,8 @@ vailabel-project = { workspace = true }
 vailabel-dataset = { workspace = true }
 vailabel-analysis = { workspace = true }
 vailabel-video = { workspace = true }
+vailabel-models = { workspace = true }
+vailabel-copilot = { workspace = true }
 diesel = { version = "2", features = ["sqlite"] }
 libsqlite3-sys = { version = "0.28", features = ["bundled"] }
 reqwest = { version = "0.12", default-features = false, features = ["blocking", "rustls-tls", "json"] }
